refactor: rename extension to xlib
